moves piano and controls to instruments folder
