+++ src/components/Instruments/Piano/DropDown.jsx
@@ -1,9 +1,9 @@
-import { pianoInstrumentType, } from '../../utils/toneUtils';
+import { pianoInstrumentType } from '../../../utils/toneUtils';
 import styles from './Piano.css';
 
 export default function Dropdown({ handleChangeType }) {
   return (
-    <select onChange={ handleChangeType } className={styles.dropDownPianoSelect}>
+    <select onChange={handleChangeType} className={styles.dropDownPianoSelect}>
       {pianoInstrumentType.map((synth, index) => (
         <option key={synth} value={synth}>
           {pianoInstrumentType[index]}
