+++ src/components/Instruments/Piano/EffectsRack.jsx
