+++ src/components/Instruments/Piano/Faders.jsx
