+++ src/components/Instruments/Piano/Oscilloscope.jsx
